backend/python-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
docx2txt==0.8
spacy==3.7.4
python-dotenv==1.0.0
pydantic==2.5.0
python-dateutil==2.8.2
numpy==1.24.3
opencv-python==4.8.1.78
aiofiles==23.2.1
